Cleaning up compiler warnings under MS VC6
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@759 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/avr/ralloc.c
@@ -1174,7 +1174,7 @@ serialRegAssign (eBBlock ** ebbs, int count)
 				/* we assign registers to it */
 				_G.regAssigned = bitVectSetBit (_G.regAssigned, sym->key);
 				if (needsPair(ic)) {
-					int regtype ;
+					short regtype ;
 					regs *preg;
 					if (sym->regType == REG_PTR) regtype = REG_PTR;
 					else if (sym->regType == REG_SCR) regtype = REG_SCR;
